skillmatch_backend/requirements.txt
absl-py==2.1.0
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
Bext==0.1.1
bleach==6.1.0
blinker==1.7.0
CacheControl==0.14.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.5
comm==0.2.1
contourpy==1.2.0
crispy-bootstrap4==2024.1
cryptography==43.0.3
csvfile==3.4.0
cycler==0.12.1
darkdetect==0.8.0
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
distro==1.9.0
Django==5.0
django-cors-headers==4.6.0
django-crispy-forms==2.3
djangorestframework==3.15.2
dm-tree==0.1.8
et-xmlfile==1.1.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.14.0
firebase-admin==6.6.0
flatbuffers==23.5.26
fonttools==4.44.0
fqdn==1.5.1
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.42
google-api-core==2.23.0
google-api-python-client==2.153.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.68.0
grpcio-status==1.68.0
h11==0.14.0
h5py==3.10.0
httpcore==1.0.4
httplib2==0.22.0
httpx==0.27.0
idna==3.5
importlib-metadata==7.0.1
iniconfig==2.0.0
ipykernel==6.29.3
ipython==8.22.1
ipywidgets==8.1.2
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
json5==0.9.22
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.4
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.13.0
jupyter_server_terminals==0.5.2
jupyterlab==4.1.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.3
jupyterlab_widgets==3.0.10
keras==3.0.5
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.3.2
MouseInfo==0.1.3
msgpack==1.1.0
namex==0.0.7
nbclient==0.9.0
nbconvert==7.16.2
nbformat==5.9.2
nest-asyncio==1.6.0
notebook==7.1.1
notebook_shim==0.2.4
numpy==1.26.1
oauthlib==3.2.2
openai==1.12.0
openpyxl==3.1.2
opt-einsum==3.3.0
overrides==7.7.0
packaging==23.2
pandas==2.2.0
pandocfilters==1.5.1
parso==0.8.3
Pillow==10.1.0
pipenv==2023.12.1
platformdirs==4.2.0
plotly==5.18.0
pluggy==1.4.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
proto-plus==1.25.0
protobuf==5.28.3
psutil==5.9.8
pure-eval==0.2.2
pyarrow==15.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAutoGUI==0.9.54
pycparser==2.21
pydantic==2.6.2
pydantic_core==2.16.3
pydeck==0.8.1b0
pydraw==2.1.3
pygame==2.5.2
PyGetWindow==0.0.9
Pygments==2.17.2
PyJWT==2.10.0
PyMsgBox==1.0.9
PyMySQL==1.1.0
pyparsing==3.1.1
PyPDF2==3.0.1
pyperclip==1.8.2
PyQt6==6.6.0
PyQt6-Fluent-Widgets==1.3.7
PyQt6-Frameless-Window==0.3.6
PyQt6-Qt6==6.6.0
PyQt6-sip==13.6.0
PyRect==0.2.0
PyScreeze==0.1.30
pytest==8.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
pytweening==1.2.0
pytz==2024.1
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.18.0
rsa==4.9
Send2Trash==1.8.2
setuptools==69.1.1
sevseg==0.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
streamlit==1.31.1
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.0rc0
tensorflow-intel==2.16.0rc0
termcolor==2.4.0
terminado==0.18.0
tinycss2==1.2.1
tk==0.1.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
tweepy==4.14.0
types-python-dateutil==2.8.19.20240106
typing_extensions==4.10.0
tzdata==2023.3
tzlocal==5.2
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.1.0
validators==0.22.0
virtualenv==20.26.1
watchdog==4.0.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
wheel==0.42.0
widgetsnbextension==4.0.10
wrapt==1.16.0
wxPython==4.2.1
zipp==3.17.0
